User Login

+++ studybud/base/views.py
@@ -1,13 +1,43 @@
 from django.shortcuts import render, redirect
+from django.contrib import messages
 from .models import Room, Topic
 from .forms import RoomForm
+from django.http import HttpResponse
 from django.db.models import Q 
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.models import User
+
+def loginPage(request): 
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        try: 
+            user = User.objects.get(username=username)
+        except: 
+            messages.error(request, 'User does not exist')
+        
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None: 
+            login(request, user)
+            return redirect('home')
+        else: 
+             messages.error(request, 'Username or password does not exist')
+
+    context = {}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutuser(request): 
+    logout(request)
+    return redirect('home')
 
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
-    rooms = Room.objects.filter(Q(topic__name__icontains=q)
-    | Q(name__icontains=q) |
+    rooms = Room.objects.filter(Q(topic__name__icontains=q) | 
+    Q(name__icontains=q) |
     Q(descriptions__icontains=q)
     )
 
@@ -22,6 +52,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='/login')
 def createRoom(request): 
     form = RoomForm()
     if request.method == 'POST': 
@@ -33,9 +64,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request, pk): 
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host: 
+        return HttpResponse('You are not allowed here!')
 
     if request.method == 'POST': 
         form = RoomForm(request.POST, instance=room)
@@ -46,8 +81,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk): 
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host: 
+        return HttpResponse('You are not allowed here!')
+
     if request.method == 'POST': 
         room.delete()
         return redirect('home')
